+++ meetup2apricot/meetup_to_apricot_event_adaptor.py
@@ -118,6 +118,10 @@ class MeetupToApricotEventAdaptor:
                     "WaitlistSettings": {
                         "WaitlistType": "Manual",
                         "InformationToCollect": "ContactInformation"
+                        },
+                    "AttendeesDisplaySettings": {
+                            "VisibleTo": "Members",
+                            "ShowPendingAttendees": False
                         }
                     }
                 }